Audio tag

+++ cron.php
@@ -26,15 +26,25 @@ foreach ($podcasts as $podcastDb) {
 
 	$podcastRssUrl = $podcastDb['url'];
 	mylog("Getting podcast info for " . $podcastDb['title']);
-	$xmlPodcast = file_get_contents($podcastRssUrl);
-	$podcast = new SimpleXMLElement($xmlPodcast);
-	$podcast = json_decode(json_encode($podcast),true); // Array
+	$xml = file_get_contents($podcastRssUrl);
+	$podcastXml = new SimpleXMLElement($xml);
+	$podcast = json_decode(json_encode($podcastXml),true); // Array
 
 	// Retrieve podcast and episodes info
 	$podcastTitle = $podcast['channel']['title'];
-	$podcastSlug = url_slug($podcastTitle);
-	$podcastImage = $podcast['channel']['image']['url'];
+	$podcastSlug = $podcastDb['slug'];
 	$episodes = $podcast['channel']['item'];
+
+	// Podcast image
+	if (isset($podcast['channel']['image']['url'])) {
+		$podcastImage = $podcast['channel']['image']['url'];
+	} elseif (isset($podcast['channel']['itunes:image']['href'])) {
+		$podcastImage = $podcastXml->channel->children("http://www.itunes.com/dtds/podcast-1.0.dtd")->image->attributes()['href'];
+	} else {
+		$podcastImage = false;
+	}
+
+	print_r($podcastImage);
 
 	if (!empty($episodes)) {
 
@@ -42,23 +52,27 @@ foreach ($podcasts as $podcastDb) {
 
 			mylog("The podcast is not parsed. Let's save the image.");
 
-			// Copy in the filesystem
-			$tmpImage = TMP_PATH . '/' . $podcastSlug;
-			copy($podcastImage, $tmpImage);
+			if ($podcastImage) {
 
-			// Get the type
-			$mime = getMimeInfoFromImage($tmpImage);
+				// Copy in the filesystem
+				$tmpImage = TMP_PATH . '/' . $podcastSlug;
+				copy($podcastImage, $tmpImage);
 
-			// Create Thumbnail
-			$imageThumbnail = new SimpleImage($tmpImage);
-			$imageThumbnail->adaptive_resize(400,400)->save($tmpImage);
+				// Get the type
+				$mime = getMimeInfoFromImage($tmpImage);
 
-			// Upload to Amazon S3
-			$uploadName = $podcastSlug.$mime['extension'];
-			$s3->putObject(S3::inputFile($tmpImage, false), AWS_S3_BUCKET, $uploadName, S3::ACL_PUBLIC_READ,array(), array('Content-Type' => $mime['contentType']));
+				// Create Thumbnail
+				$imageThumbnail = new SimpleImage($tmpImage);
+				$imageThumbnail->adaptive_resize(400,400)->save($tmpImage);
 
-			// Delete the tmp image
-			unlink($tmpImage);
+				// Upload to Amazon S3
+				$uploadName = $podcastSlug.$mime['extension'];
+				$s3->putObject(S3::inputFile($tmpImage, false), AWS_S3_BUCKET, $uploadName, S3::ACL_PUBLIC_READ,array(), array('Content-Type' => $mime['contentType']));
+
+				// Delete the tmp image
+				unlink($tmpImage);
+
+			}
 
 			// DB mark podcast as parsed
 			$db->rawQuery("UPDATE podcasts SET parsed = 1 WHERE id = ?", array($podcastDb['id']));
@@ -80,46 +94,11 @@ foreach ($podcasts as $podcastDb) {
 			$db->where('slug',$slug);
 			$exists = $db->getOne('episodes');
 
-			// If not exists, we sox it and save it to DB
+			// If not exists, we save it to DB
 			// Get mp3
 			if (!$exists) {
 
-				mylog($title . " is not parsed yet, let's parse it.");
-
-				$tmpMp3 = TMP_PATH . '/' . $slug . '.mp3';
-
-				mylog("Retrieving mp3 file...");
-				copy($urlMp3, $tmpMp3);
-
-				// Once copied, we create the slowed versions
-				$versions = array(70,80,90,120);
-				foreach ($versions as $version) {
-
-					try {
-
-						// Run sox command
-						mylog("Generating version: " . $version . ". Please wait...");
-						$tmpMp3Slowed = TMP_PATH . '/' . $slug . "_" . $version . '.mp3';
-						exec('sox ' . $tmpMp3 . ' ' . $tmpMp3Slowed . ' tempo ' . $version/100);
-
-						mylog("Uploading to Amazon S3: " . $version . ". Please wait...");
-						$uploadName = $podcastSlug. '/' . $slug . "_" . $version . '.mp3';
-						$s3->putObject(S3::inputFile($tmpMp3Slowed, false), AWS_S3_BUCKET, $uploadName, S3::ACL_PUBLIC_READ,array(), array('Content-Type' => 'audio/mpeg'));
-
-						// Delete the tmp mp3 version
-						unlink($tmpMp3Slowed);
-
-					} catch (Exception $e) {
-
-						//file_put_contents($filename, $data);
-						mylog('Error: ' . $e->getMessage());
-
-					}
-
-				}
-
-				// Delete the tmp mp3
-				unlink($tmpMp3);
+				mylog($title . " is not parsed yet, let's save it.");
 
 				// Save info in DB
 				mylog("Saving info in DB");
